requirements.txt.txt
python-telegram-bot==20.6 
pandas 
openpyxl